experiment/requirements.txt
nebula3-python==3.8.3
libclang
clang==16.0.1
loguru>=0.7.2
